initial update to support llvm 17
