+++ tsc/include/TypeScript/TypeScriptDialect.td
@@ -24,8 +24,6 @@ def TypeScript_Dialect : Dialect {
 
     let useDefaultTypePrinterParser = 1;
     //let useDefaultAttributePrinterParser = 1;
-
-    let useFoldAPI = kEmitFoldAdaptorFolder;
 }
 
 //===----------------------------------------------------------------------===//
