+++ tsc/test/tester/test-runner.cpp
@@ -8,7 +8,7 @@
 #ifdef WIN32
 #define TYPESCRIPT_LIB "TypeScriptAsyncRuntime.lib "
 #define LLVM_LIBS "LLVMSupport.lib "
-#define LIBS "msvcrt"_D_".lib ucrt"_D_".lib "
+#define LIBS "msvcrt" _D_ ".lib ucrt" _D_ ".lib "
 #else
 // for Ubuntu 20.04 add -ldl and optionally -rdynamic 
 #define LIBS "-frtti -fexceptions -lstdc++ -lrt -ldl -lpthread -lm -ltinfo"
